+++ assets/js/phoenix_live_view/live_socket.js
@@ -95,6 +95,7 @@ import {
   PHX_THROTTLE,
   PHX_TRACK_UPLOADS,
   PHX_SESSION,
+  PHX_FEEDBACK_FOR,
   RELOAD_JITTER_MIN,
   RELOAD_JITTER_MAX,
 } from "./constants"
@@ -855,6 +856,15 @@ export default class LiveSocket {
         })
       }, false)
     }
+    this.on("reset", (e) => {
+      let form = e.target
+      DOM.resetForm(form, this.binding(PHX_FEEDBACK_FOR))
+      let input = Array.from(form.elements).find(el => el.type === "reset")
+      // wait until next tick to get updated input value
+      window.requestAnimationFrame(() => {
+        input.dispatchEvent(new Event("input", {bubbles: true, cancelable: false}))
+      })
+    })
   }
 
   debounce(el, event, eventType, callback){