+++ lib/phoenix_component.ex
@@ -153,7 +153,7 @@ defmodule Phoenix.Component do
   But what if you need your function components to support dynamic attributes, such as common HTML
   attributes to mix into a component's container?
 
-  ### Global Attributes
+  ## Global attributes
 
   Global attributes are a set of attributes that a function component can accept when it
   declares an attribute of type `:global`. By default, the set of attributes accepted are those
@@ -231,26 +231,24 @@ defmodule Phoenix.Component do
   end
   ```
 
-  The `:include` option is useful to apply global additions on a case-by-case basis, but
-  sometimes you want attributes to be available to all globals you provide, such
-  as when using frameworks that use attribute prefixes, like Alpine.js's `x-on:click`.
-  For these cases, custom global attribute prefixes can be provided, which we'll outline
-  next.
+  The `:include` option is useful to apply global additions on a case-by-case basis,
+  but sometimes you want to extend existing components with new global attributes,
+  such as Alpine.js' `x-` prefixes, which we'll outline next.
 
-  ### Custom Global Attribute Prefixes
+  ### Custom global attribute prefixes
 
   You can extend the set of global attributes by providing a list of attribute prefixes to
   `use Phoenix.Component`. Like the default attributes common to all HTML elements,
   any number of attributes that start with a global prefix will be accepted by function
-  components defined in this module. By default, the following prefixes are supported:
+  components invoked by the current module. By default, the following prefixes are supported:
   `phx-`, `aria-`, and `data-`. For example, to support the `x-` prefix used by
   [Alpine.js](https://alpinejs.dev/), you can pass the `:global_prefixes` option to
   `use Phoenix.Component`:
 
       use Phoenix.Component, global_prefixes: ~w(x-)
 
-  Now all function components defined in this module will accept any number of attributes prefixed
-  with `x-`, in addition to the default global prefixes.
+  Now all function components invoked by this module will accept any number of attributes
+  prefixed with `x-`, in addition to the default global prefixes.
 
   You can learn more about attributes by reading the documentation for `attr/3`.
 
@@ -1251,8 +1249,8 @@ defmodule Phoenix.Component do
 
   def assign(%{__changed__: changed} = assigns, key, value) do
     case assigns do
-      # force assign the key if the attribute was declared with default
-      %{^key => ^value, __defaults__: %{^key => _}} ->
+      # force assign the key if the attribute was given with matching value
+      %{^key => ^value, __given__: given} when not is_map_key(given, key) ->
         Phoenix.LiveView.Utils.force_assign(assigns, changed, key, value)
 
       %{^key => ^value} ->
@@ -1372,22 +1370,31 @@ defmodule Phoenix.Component do
   according to `Phoenix.HTML.FormData`.
 
   This is commonly used to convert a map or an Ecto changeset
-  into a form to be given to the `form/1` component. For example,
-  if you want to create a form based on `handle_event` parameters,
+  into a form to be given to the `form/1` component.
+
+  ## Creating a form from params
+
+  If you want to create a form based on `handle_event` parameters,
   you could do:
 
       def handle_event("submitted", params, socket) do
         {:noreply, assign(socket, form: to_form(params))}
       end
 
-  However, most typically, we specify a name to nest the parameters:
+  When you pass a map to `to_form/1`, it assumes said map contains
+  the form parameters, which are expected to have string keys.
+
+  You can also specify a name to nest the parameters:
 
       def handle_event("submitted", %{"user" => user_params}, socket) do
         {:noreply, assign(socket, form: to_form(user_params, as: :user))}
       end
 
-  When using changesets, the name `:as` is automatically retrieved
-  from the schema. For example, if you have a user schema:
+  ## Creating a form from changesets
+
+  When using changesets, the underlying data, form parameters, and
+  errors are retrieved from it. The `:as` option is automatically
+  computed too. For example, if you have a user schema:
 
       defmodule MyApp.Users.User do
         use Ecto.Schema
@@ -1403,15 +1410,8 @@ defmodule Phoenix.Component do
       |> Ecto.Changeset.change()
       |> to_form()
 
-  Phoenix will take care of getting all of the relevant information
-  from the changeset, including errors, data, and names for you.
-  In this case, the parameters will be available under
-  `%{"post" => post_params}`.
-
-  If an existing `Phoenix.HTML.Form` struct is given, the
-  options below will override its existing values if given.
-  Then the remaining options are merged with the existing
-  form options.
+  In this case, once the form is submitted, the parameters will
+  be available under `%{"post" => post_params}`.
 
   ## Options
 
@@ -1422,8 +1422,13 @@ defmodule Phoenix.Component do
   converted to forms. For example, a map accepts `:errors`
   to list errors, but such option is not accepted by
   changesets.
+
+  If an existing `Phoenix.HTML.Form` struct is given, the
+  options below will override its existing values if given.
+  Then the remaining options are merged with the existing
+  form options.
   """
-  def to_form(data, options \\ [])
+  def to_form(data_or_params, options \\ [])
 
   def to_form(%Phoenix.HTML.Form{} = data, options) do
     {name, id} =
@@ -1821,7 +1826,7 @@ defmodule Phoenix.Component do
       end
   '''
   @doc type: :macro
-  defmacro attr(name, type, opts \\ []) when is_atom(name) and is_list(opts) do
+  defmacro attr(name, type, opts \\ []) do
     quote bind_quoted: [name: name, type: type, opts: opts] do
       Phoenix.Component.Declarative.__attr__!(
         __MODULE__,
@@ -1954,14 +1959,12 @@ defmodule Phoenix.Component do
   and generates the relevant form tags. It can be used either inside LiveView
   or outside.
 
-  [INSERT LVATTRDOCS]
-
   ## Examples: inside LiveView
 
-  Inside LiveViews, the `:for` attribute is generally a form struct created
-  with the `to_form/1` function. `to_form/1` expects either a map or an
-  [`Ecto.Changeset`](https://hexdocs.pm/ecto/Ecto.Changeset.html) as the
-  source of data.
+  Inside LiveViews, the `for={...}` attribute is generally a form struct
+  created with the `to_form/1` function. `to_form/1` expects either a map
+  or an [`Ecto.Changeset`](https://hexdocs.pm/ecto/Ecto.Changeset.html)
+  as the source of data.
 
   For example, you may use the parameters received in a
   `c:Phoenix.LiveView.handle_event/3` callback to create an Ecto changeset
@@ -2084,6 +2087,8 @@ defmodule Phoenix.Component do
   party applications. If this behaviour is problematic, you can generate
   a non-host specific token with `Plug.CSRFProtection.get_csrf_token/0` and
   pass it to the form generator via the `:csrf_token` option.
+
+  [INSERT LVATTRDOCS]
   '''
   @doc type: :component
   attr.(:for, :any, required: true, doc: "An existing form or the form source data.")
@@ -2130,6 +2135,14 @@ defmodule Phoenix.Component do
     """
   )
 
+  attr.(:errors, :list,
+    doc: """
+    Use this to manually pass a keyword list of errors to the form, such as `@errors`.
+    This option is useful when a regular map is given as the form source and it will
+    make the errors available under `f.errors`.
+    """
+  )
+
   attr.(:rest, :global,
     include: ~w(autocomplete name rel enctype novalidate target),
     doc: "Additional HTML attributes to add to the form tag."
@@ -2138,10 +2151,10 @@ defmodule Phoenix.Component do
   slot.(:inner_block, required: true, doc: "The content rendered inside of the form tag.")
 
   def form(assigns) do
-    # Extract options and then to the same call as form_for
     action = assigns[:action]
 
     form_for =
+      # We require for={...} to be given but we automatically handle nils for convenience
       case assigns[:for] do
         nil -> %{}
         other -> other