+++ lib/phoenix_live_view/html_formatter.ex
@@ -430,7 +430,7 @@ defmodule Phoenix.LiveView.HTMLFormatter do
   end
 
   defp to_tree(
-         [{:text, text, %{context: [:comment_end]}} | tokens],
+         [{:text, text, %{context: [:comment_end | _rest]}} | tokens],
          buffer,
          [{:comment, start_text, upper_buffer} | stack],
          source
@@ -573,7 +573,7 @@ defmodule Phoenix.LiveView.HTMLFormatter do
   #   <!-- Modal content -->
   #   <%= render_slot(@inner_block) %>
   #
-  # Thefore the case above will stay as is. Otherwise it would put them in the
+  # Therefore the case above will stay as is. Otherwise it would put them in the
   # same line.
   defp line_html_comment?(text) do
     trimmed_text = String.trim(text)